+++ packages/node-modules-inspector/src/app/utils/maintainer-action-templates.ts
@@ -52,12 +52,12 @@ export function buildAgentPrompt(item: MaintainerActionItem): string {
   const newOnlyExample = toMajor !== undefined ? `\`^${toMajor}.0.0\`` : `the new major`
 
   const intent = item.depType === 'peer'
-    ? `\`${item.consumer.name}\` declares \`${item.depName}\` as a peer dependency — it signals compatibility to consumers. The goal is to broaden that compatibility, not necessarily to drop the old major.`
-    : `\`${item.consumer.name}\` consumes \`${item.depName}\` directly. The goal is to upgrade to ${toLabel}.`
+    ? `\`${item.depName}\` is declared as a peer dependency — it signals compatibility to downstream consumers. The goal is to broaden that compatibility, not necessarily to drop the old major.`
+    : `\`${item.depName}\` is a direct dependency. The goal is to upgrade to ${toLabel}.`
 
   const compatBullet = item.depType === 'peer'
-    ? `- If no breaking change affects \`${item.consumer.name}\`, widen \`${block}.${item.depName}\` to a union (e.g. ${unionExample}) so consumers on either major are still supported.`
-    : `- If no breaking change affects \`${item.consumer.name}\`, update \`${block}.${item.depName}\` to allow ${toLabel}. A union (${unionExample}) keeps backward compat; a clean bump to ${newOnlyExample} is simpler if older support is not needed.`
+    ? `- If no breaking change affects this package, widen \`${block}.${item.depName}\` to a union (e.g. ${unionExample}) so consumers on either major are still supported.`
+    : `- If no breaking change affects this package, update \`${block}.${item.depName}\` to allow ${toLabel}. A union (${unionExample}) keeps backward compat; a clean bump to ${newOnlyExample} is simpler if older support is not needed.`
 
   const lines: string[] = [
     `Update \`${item.depName}\` in \`${item.consumer.name}\` from ${fromLabel} to ${toLabel} (specifically \`${item.depName}@${item.installedHighestVersion}\`).`,
@@ -73,7 +73,7 @@ export function buildAgentPrompt(item: MaintainerActionItem): string {
     '',
     `- Review \`${item.depName}\`'s changelog between ${fromLabel} and ${toLabel} to identify breaking changes.`,
     compatBullet,
-    `- If a breaking change does affect \`${item.consumer.name}\`, do not change the declaration unilaterally — pause and ask the user how to proceed (e.g. drop the old major and bump, fork compatibility, etc.).`,
+    `- If a breaking change does affect this package, do not change the declaration unilaterally — pause and ask the user how to proceed (e.g. drop the old major and bump, fork compatibility, etc.).`,
     `- Verify the update by running the test suite, linting, and type checks.`,
   )
 
@@ -89,29 +89,27 @@ export function buildAgentPromptAll(consumer: PackageNode, items: MaintainerActi
   const lines: string[] = [
     `Update ${items.length} ${mix} dependencies in \`${consumer.name}\`:`,
     '',
-    `| Dep | Type | From | Target |`,
-    `| --- | --- | --- | --- |`,
     ...items.map((i) => {
       const block = i.depType === 'peer' ? 'peerDependencies' : 'dependencies'
       const fromMajor = rangeHighestMajor(i.declaredRange)
-      const from = fromMajor !== undefined ? `v${fromMajor} (\`${i.declaredRange}\`)` : `\`${i.declaredRange}\``
+      const fromLabel = fromMajor !== undefined ? `v${fromMajor} (\`${i.declaredRange}\`)` : `\`${i.declaredRange}\``
       const toMajor = safeMajor(i.installedHighestVersion)
-      const to = toMajor !== undefined ? `v${toMajor} (\`${i.depName}@${i.installedHighestVersion}\`)` : `\`${i.depName}@${i.installedHighestVersion}\``
-      return `| \`${i.depName}\` | \`${block}\` | ${from} | ${to} |`
+      const toLabel = toMajor !== undefined ? `v${toMajor} (\`${i.depName}@${i.installedHighestVersion}\`)` : `\`${i.depName}@${i.installedHighestVersion}\``
+      return `- \`${block}.${i.depName}\`: ${fromLabel} → ${toLabel}`
     }),
   ]
 
   if (consumerData.repository)
-    lines.push('', `\`${consumer.name}\` repository: ${consumerData.repository}`)
+    lines.push('', `Repository: ${consumerData.repository}`)
 
   lines.push(
     '',
     `For each dep:`,
     `- Review its changelog between the current and target majors to identify breaking changes.`,
-    `- If no breaking change affects \`${consumer.name}\`:`,
+    `- If no breaking change affects this package:`,
     `  - For peer dependencies, widen the range to a union of the old and new majors.`,
     `  - For direct dependencies, update the range to accept the new major (union for backward compat, or a clean bump if older support is not needed).`,
-    `- If a breaking change does affect \`${consumer.name}\`, pause on that dep and ask the user how to proceed.`,
+    `- If a breaking change does affect this package, pause on that dep and ask the user how to proceed.`,
     '',
     `After all updates, verify with the test suite, linting, and type checks.`,
   )
